# [add] added hasdata parameter to RAI_TensorCreate, which calls either…