--- a/src/tensor.h
+++ b/src/tensor.h
@@ -9,7 +9,7 @@
 extern RedisModuleType *RedisAI_TensorType;
 
 int RAI_TensorInit(RedisModuleCtx* ctx);
-RAI_Tensor* RAI_TensorCreate(const char* dataTypeStr, long long* dims, int ndims);
+RAI_Tensor* RAI_TensorCreate(const char* dataTypeStr, long long* dims, int ndims, int hasdata);
 RAI_Tensor* RAI_TensorCreateFromDLTensor(DLManagedTensor* dl_tensor);
 size_t RAI_TensorLength(RAI_Tensor* t);
 size_t RAI_TensorGetDataSize(const char* dataTypeStr);